requirements.txt
langchain_core
langchain
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
pydantic-settings
langchain-ollama
langchain-classic
arxiv
wikipedia
langchain_text_splitters
langchainhub